+++ NEWS.md
@@ -4,3 +4,5 @@
 ## v0.5.0
 
 - Support Typst v0.13.0
+- Implement `read(::TypstCommand)` and `read(::TypstCommand, ::Type{String})`
+    - Closes [Support read for TypstCommand #14](https://github.com/jakobjpeters/Typstry.jl/issues/14)